+++ server/plugins/inputs/ban.ts
@@ -1,6 +1,7 @@
 import {ChanType} from "../../models/chan";
-import Msg, {MessageType} from "../../models/msg";
+import Msg from "../../models/msg";
 import {PluginInputHandler} from "./index";
+import {MessageType} from "../../../shared/types/msg";
 
 const commands = ["ban", "unban", "banlist", "kickban"];
