+++ server/plugins/inputs/mute.ts
@@ -2,9 +2,10 @@ import Chan from "../../models/chan";
 import Network from "../../models/network";
 import {PluginInputHandler} from "./index";
 
-import Msg, {MessageType} from "../../models/msg";
+import Msg from "../../models/msg";
 
 import Client from "../../client";
+import {MessageType} from "../../../shared/types/msg";
 
 const commands = ["mute", "unmute"];
 const allowDisconnected = true;
