+++ server/plugins/inputs/part.ts
@@ -1,8 +1,9 @@
 import {PluginInputHandler} from "./index";
 
-import Msg, {MessageType} from "../../models/msg";
+import Msg from "../../models/msg";
 import Config from "../../config";
 import {ChanType, ChanState} from "../../models/chan";
+import {MessageType} from "../../../shared/types/msg";
 
 const commands = ["close", "leave", "part"];
 const allowDisconnected = true;
