+++ server/plugins/inputs/rejoin.ts
@@ -1,7 +1,8 @@
 import {PluginInputHandler} from "./index";
 
-import Msg, {MessageType} from "../../models/msg";
+import Msg from "../../models/msg";
 import {ChanType} from "../../models/chan";
+import {MessageType} from "../../../shared/types/msg";
 
 const commands = ["cycle", "rejoin"];
