+++ server/plugins/irc-events/connection.ts
@@ -3,10 +3,11 @@ import _ from "lodash";
 import {IrcEventHandler} from "../../client";
 
 import log from "../../log";
-import Msg, {MessageType} from "../../models/msg";
+import Msg from "../../models/msg";
 import Helper from "../../helper";
 import Config from "../../config";
 import {ChanType, ChanState} from "../../models/chan";
+import {MessageType} from "../../../shared/types/msg";
 
 export default <IrcEventHandler>function (irc, network) {
 	const client = this;
