+++ server/plugins/irc-events/help.ts
@@ -1,5 +1,6 @@
-import Msg, {MessageType} from "../../models/msg";
+import Msg from "../../models/msg";
 import {IrcEventHandler} from "../../client";
+import {MessageType} from "../../../shared/types/msg";
 
 export default <IrcEventHandler>function (irc, network) {
 	const client = this;
