+++ server/plugins/irc-events/kick.ts
@@ -1,8 +1,9 @@
 import {IrcEventHandler} from "../../client";
 import {ChanState} from "../../models/chan";
 
-import Msg, {MessageType} from "../../models/msg";
+import Msg from "../../models/msg";
 import User from "../../models/user";
+import {MessageType} from "../../../shared/types/msg";
 
 export default <IrcEventHandler>function (irc, network) {
 	const client = this;
