+++ server/plugins/irc-events/message.ts
@@ -1,10 +1,11 @@
-import Msg, {MessageType} from "../../models/msg";
+import Msg from "../../models/msg";
 import LinkPrefetch from "./link";
 import {cleanIrcMessage} from "../../../shared/irc";
 import Helper from "../../helper";
 import {IrcEventHandler} from "../../client";
 import Chan, {ChanType} from "../../models/chan";
 import User from "../../models/user";
+import {MessageType} from "../../../shared/types/msg";
 
 const nickRegExp = /(?:\x03[0-9]{1,2}(?:,[0-9]{1,2})?)?([\w[\]\\`^{|}-]+)/g;
