+++ server/plugins/irc-events/modelist.ts
@@ -1,7 +1,8 @@
 import {IrcEventHandler} from "../../client";
 import {SpecialChanType, ChanType} from "../../models/chan";
 
-import Msg, {MessageType} from "../../models/msg";
+import Msg from "../../models/msg";
+import {MessageType} from "../../../shared/types/msg";
 
 export default <IrcEventHandler>function (irc, network) {
 	const client = this;
