+++ server/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../tsconfig.base.json" /* Path to base configuration file to inherit from. Requires TypeScript version 2.1 or later. */,
   "include": [
-    "**/*",
+    ".",
     "../shared/"
   ] /* Specifies a list of glob patterns that match files to be included in compilation. If no 'files' or 'include' property is present in a tsconfig.json, the compiler defaults to including all files in the containing directory and subdirectories except those specified by 'exclude'. Requires TypeScript version 2.0 or later. */,
   "files": [
