+++ server/types/index.d.ts
@@ -1,2 +1,1 @@
 import "./modules";
-import "./socket-events";
