+++ shared/types/socket-events.d.ts
